moved markdown posts to components folder
